Setup for CCR
This build script will use the modulefile templates for CCR for GCC
and Intel MPI and setup modulefiles that we can load but that also
set our environment variables. Internally, we're just loading CCR's
modulefiles. We'll need to update the build script as we move either
compiler, MPI stack, or both.