Hardcode texlive builder to 2017
CentOS6 can't deal with 2018 texlive and everything before 2018
has been yanked off of all the repos, so we have to use the historical
repo for texlive 2017 until we get a newer system. Should probably
do this over a weekend when there's less traffic. Their server seems
to tend to disconnect if it gets busy. This took about 4 hours for me.